python/airflow-docker-basics/requirements.txt
apache-airflow==2.8.4
